ui/rsl: add service class filter to capacity status
